Renamed GNPSFileMappingLoader
